Pull getPackages into util and use in checkUpdatedPackages
