+++ lib/utils/checkUpdatedPackages.js
@@ -1,9 +1,9 @@
 var getPackageLocation = require("./getPackageLocation");
 var getPackageConfig   = require("./getPackageConfig");
+var getPackages        = require("./getPackages");
 var progressBar        = require("./progressBar");
 var execSync           = require("./execSync");
 var chalk              = require("chalk");
-var fs                 = require("fs");
 
 module.exports = function checkUpdatedPackages(packagesLoc) {
   var changedPackages = [];
@@ -12,8 +12,9 @@ module.exports = function checkUpdatedPackages(packagesLoc) {
 
   console.log("Checking packages...");
 
-  var packageNames = fs.readdirSync(packagesLoc).filter(function (name) {
-    return name[0] !== "." && fs.statSync(packagesLoc + "/" + name).isDirectory();
+  var packages = getPackages(packagesLoc);
+  var packageNames = packages.map(function(pkg) {
+    return pkg.name;
   });
 
   var tick = progressBar(packageNames.length);
